+++ src/episodes/tests/episodes.experiments.test.ts
@@ -1,16 +1,16 @@
 import { assert, assertEquals } from '@std/assert';
 import { EpisodesRepository } from '../repository/episodes.repository.ts';
 import { toCanonicalEpisode } from '../episodes.types.ts';
-import { Features } from '../../../common/types/core.ts';
+import { Features } from '../../common/types/core.ts';
 import type { EpisodeCollection } from '../collection/episode.collection.ts';
 import type { EpisodeDocument } from '../store/types.ts';
-import { AppFeatures } from '../../../common/experiment/types.ts';
+import { AppFeatures } from '../../common/experiment/types.ts';
 // We'll stub remotes via fetch helpers
 import type { TheXemDataModel } from '../../service/thexem/remote/types.ts';
 import type { SkyhookShow } from '../../service/skyhook/types.ts';
-import { setEnvScoped } from '../../../common/testing/env.ts';
-import { json, onGet, stubFetch } from '../../../common/testing/net.ts';
-import { Instant } from '../../../common/helpers/date.ts';
+import { setEnvScoped } from '../../common/testing/env.ts';
+import { json, onGet, stubFetch } from '../../common/testing/net.ts';
+import { Instant } from '../../common/helpers/date.ts';
 import { SeriesRelationId } from '../../service/arm/index.ts';
 
 // Minimal in-memory EpisodeCollection
