+++ src/episodes/tests/episodes.test.ts
@@ -7,11 +7,11 @@ import {
 import { toCanonicalEpisode } from '../episodes.types.ts';
 import { EpisodesRepository } from '../repository/episodes.repository.ts';
 import { clampLimit } from '../episodes.params.ts';
-import { Features } from '../../../common/types/core.ts';
+import { Features } from '../../common/types/core.ts';
 import type { EpisodeCollection } from '../collection/episode.collection.ts';
 import type { EpisodeDocument } from '../store/types.ts';
-import { AppFeatures } from '../../../common/experiment/types.ts';
-import { Instant } from '../../../common/helpers/date.ts';
+import { AppFeatures } from '../../common/experiment/types.ts';
+import { Instant } from '../../common/helpers/date.ts';
 
 // Minimal in-memory collection for repository DI in tests
