+++ src/series/repository/series.repository.ts
@@ -1,19 +1,19 @@
 import { IResponse } from '../../common/types/response.ts';
-import { getAniListRelationId } from '../service/arm/index.ts';
-import { getJikanAnime } from '../service/jikan/index.ts';
-import { getNotifyAnime } from '../service/notify/index.ts';
-import { getSkyhookShow } from '../service/skyhook/index.ts';
-import { getThemesForAnime } from '../service/theme/index.ts';
-import { getTmdbShow } from '../service/tmdb/index.ts';
-import { getTraktShow } from '../service/trakt/index.ts';
+import { getAniListRelationId } from '../../service/arm/index.ts';
+import { getJikanAnime } from '../../service/jikan/index.ts';
+import { getNotifyAnime } from '../../service/notify/index.ts';
+import { getSkyhookShow } from '../../service/skyhook/index.ts';
+import { getThemesForAnime } from '../../service/theme/index.ts';
+import { getTmdbShow } from '../../service/tmdb/index.ts';
+import { getTraktShow } from '../../service/trakt/index.ts';
 import { seriesTransform } from '../transformer/series.transformer.ts';
 import { MediaEntity } from '../types.ts';
 import { isManga } from '../utils/index.ts';
 import LocalSource from '../local/series.local.source.ts';
-import { Theme } from '../service/theme/transformer/types.ts';
-import { SkyhookShow } from '../service/skyhook/types.ts';
-import { Show } from '../service/trakt/transformer/types.ts';
-import { TmdbShow } from '../service/tmdb/types.ts';
+import { Theme } from '../../service/theme/transformer/types.ts';
+import { SkyhookShow } from '../../service/skyhook/types.ts';
+import { Show } from '../../service/trakt/transformer/types.ts';
+import { TmdbShow } from '../../service/tmdb/types.ts';
 import { currentDate, isOlderThan } from '../../common/core/utils.ts';
 import { MediaParamId } from '../local/index.ts';
