+++ src/series/transformer/season.transformer.ts
@@ -1,7 +1,7 @@
 import { toInstant } from '../../common/helpers/date.ts';
-import { provider } from '../service/tmdb/transformer/index.ts';
-import { TmdbCrew, TmdbImage } from '../service/tmdb/types.ts';
-import { ImageProviderType } from '../service/tmdb/utils/image-provider.ts';
+import { provider } from '../../service/tmdb/transformer/index.ts';
+import { TmdbCrew, TmdbImage } from '../../service/tmdb/types.ts';
+import { ImageProviderType } from '../../service/tmdb/utils/image-provider.ts';
 import {
   SeriesEpisode,
   SeriesEpisodeCrew,
