+++ src/series/utils/series.utils.ts
@@ -1,5 +1,5 @@
-import { MalType } from '../service/jikan/remote/enums.ts';
-import { Format } from '../service/notify/transformer/enums.ts';
+import { MalType } from '../../service/jikan/remote/enums.ts';
+import { Format } from '../../service/notify/transformer/enums.ts';
 
 export const isMovie = (format?: Format): boolean => format == Format.MOVIE;
