+++ src/service/arm/transformer/arm.transformer.ts
@@ -1,6 +1,6 @@
 import { ArmModel } from '../remote/types.ts';
 import { SeriesRelationId } from '../types.ts';
-import { Transform } from '../../../../common/transformer/types.ts';
+import { Transform } from '../../../common/transformer/types.ts';
 
 export const transform: Transform<ArmModel, SeriesRelationId> = (
   sourceData,
