+++ src/service/notify/remote/notify.remote.ts
@@ -1,6 +1,6 @@
-import { env } from '../../../../common/core/env.ts';
-import { request } from '../../../../common/core/request.ts';
-import { Service } from '../../../../common/types/state.ts';
+import { env } from '../../../common/core/env.ts';
+import { request } from '../../../common/core/request.ts';
+import { Service } from '../../../common/types/state.ts';
 import { AnimeModel, EpisodeIdModel, EpisodeModel } from './types.ts';
 
 const getService = (): Service => ({
