+++ src/service/notify/transformer/notify.transformer.ts
@@ -1,5 +1,5 @@
-import { toInstant } from '../../../../common/helpers/index.ts';
-import { Transform } from '../../../../common/transformer/types.ts';
+import { toInstant } from '../../../common/helpers/index.ts';
+import { Transform } from '../../../common/transformer/types.ts';
 import {
   EpisodeModel,
   ImageModel,
