+++ src/service/skyhook/remote/skyhook.remote.ts
@@ -1,6 +1,6 @@
-import { env } from '../../../../common/core/env.ts';
-import { request } from '../../../../common/core/request.ts';
-import { Service } from '../../../../common/types/state.ts';
+import { env } from '../../../common/core/env.ts';
+import { request } from '../../../common/core/request.ts';
+import { Service } from '../../../common/types/state.ts';
 import { SkyhookModel } from './types.ts';
 
 const getService = (): Service => ({
