+++ src/service/thexem/thexem.service.test.ts
@@ -5,8 +5,8 @@ import {
   clearTheXemCache,
   getTheXemMappingsByTvdb,
 } from './thexem.service.ts';
-import { setEnvScoped } from '../../../common/testing/env.ts';
-import { json, onGet, stubFetch } from '../../../common/testing/net.ts';
+import { setEnvScoped } from '../../common/testing/env.ts';
+import { json, onGet, stubFetch } from '../../common/testing/net.ts';
 import { TheXem } from './types.ts';
 
 Deno.test('buildTvdbAbsoluteMap builds a map from tvdb.absolute to absolute', () => {
