+++ src/service/trakt/remote/trakt.remote.ts
@@ -1,6 +1,6 @@
-import { env } from '../../../../common/core/env.ts';
-import { defaults, request } from '../../../../common/core/request.ts';
-import { Service } from '../../../../common/types/state.ts';
+import { env } from '../../../common/core/env.ts';
+import { defaults, request } from '../../../common/core/request.ts';
+import { Service } from '../../../common/types/state.ts';
 import { ShowModel } from './types.ts';
 
 const getService = (): Service => ({
