+++ src/service/trakt/trakt.service.ts
@@ -1,4 +1,4 @@
-import { logger } from '../../../common/core/logger.ts';
+import { logger } from '../../common/core/logger.ts';
 import { transform } from './transformer/index.ts';
 import { getTraktShowByIdOrSlug } from './remote/index.ts';
 import { TraktShow } from './types.ts';
