+++ src/service/trakt/transformer/types.ts
@@ -1,4 +1,4 @@
-import { Instant } from '../../../../common/helpers/date.ts';
+import { Instant } from '../../../common/helpers/date.ts';
 
 interface MediaId {
   trakt: number;
